Track tag of spawning bio atom
Began work on tracking organism lineages. Each bio atom now has a spawner attribute which is assigned the value of the tag associated with the parent which spawned it. Need to double check that memory is being deallocated correctly. The included lammps git repo does have a modified dump-vtk, but that should be made into a dump bio/vtk, similar to what is being done with hdf5.